+++ src/types/struct.test.ts
@@ -33,7 +33,7 @@ test('test struct with reference', () => {
         let (input, file_number) = be_u16(input)?;
         let (input, record_number) = be_u16(input)?;
         let (input, record_len) = be_u16(input)?;
-        let (input, record_data) = take(record_len)(input)?;
+        let (input, record_data) = take(record_len as usize)(input)?;
         Ok((
             input,
             WriteFileRecordSubRequest {
